+++ packages/generator-adp/src/app/questions/configuration.ts
@@ -1,5 +1,4 @@
 import {
-    UI5VersionInfo,
     FlexLayer,
     getConfiguredProvider,
     getEndpointNames,
@@ -10,14 +9,23 @@ import {
     SourceManifest,
     isAppSupported,
     isSyncLoadedView,
-    isV4Application
+    isV4Application,
+    getRelevantVersions,
+    fetchPublicVersions,
+    checkSystemVersionPattern
 } from '@sap-ux/adp-tooling';
 import type { ToolsLogger } from '@sap-ux/logger';
 import type { Manifest } from '@sap-ux/project-access';
 import { validateEmptyString } from '@sap-ux/project-input-validator';
 import { isAxiosError, type AbapServiceProvider } from '@sap-ux/axios-extension';
 import type { InputQuestion, ListQuestion, PasswordQuestion, YUIQuestion } from '@sap-ux/inquirer-common';
-import type { ConfigAnswers, FlexUISupportedSystem, SourceApplication, SystemLookup } from '@sap-ux/adp-tooling';
+import type {
+    ConfigAnswers,
+    FlexUISupportedSystem,
+    SourceApplication,
+    SystemLookup,
+    UI5Version
+} from '@sap-ux/adp-tooling';
 
 import type {
     ApplicationPromptOptions,
@@ -35,7 +43,6 @@ import { getAppAdditionalMessages, getSystemAdditionalMessages } from './helper/
 
 /**
  * A stateful prompter class that creates configuration questions.
- * This class accepts the needed dependencies and keeps track of state (e.g. the ApplicationManager instance).
  * It exposes a single public method {@link getPrompts} to retrieve the configuration questions.
  */
 export class ConfigPrompter {
@@ -92,9 +99,39 @@ export class ConfigPrompter {
      */
     private isPartiallySupported = false;
     /**
-     * UI5 version manager for handling version-related validations.
+     * UI5 versions in string format.
      */
-    private readonly ui5Info: UI5VersionInfo;
+    private ui5Versions: string[];
+    /**
+     * Publicly available UI5 versions.
+     */
+    private publicVersions: UI5Version;
+    /**
+     * System UI5 version.
+     */
+    private systemVersion: string | undefined;
+
+    /**
+     * Returns the needed ui5 properties from calling the CDN.
+     *
+     * @returns Object with properties related to ui5.
+     */
+    public get ui5(): { ui5Versions: string[]; systemVersion: string | undefined; publicVersions: UI5Version } {
+        return {
+            ui5Versions: this.ui5Versions,
+            systemVersion: this.systemVersion,
+            publicVersions: this.publicVersions
+        };
+    }
+
+    /**
+     * Returns flag indicating if the project is a cloud project.
+     *
+     * @returns Whether system is cloud-ready.
+     */
+    public get isCloud(): boolean {
+        return !!this.isCloudProject;
+    }
 
     /**
      * Returns the configured abap provider instance.
@@ -110,8 +147,8 @@ export class ConfigPrompter {
      *
      * @returns Application manifest.
      */
-    public get manifest(): Manifest {
-        return this.manifest;
+    public get manifest(): Manifest | undefined {
+        return this.appManifest;
     }
 
     /**
@@ -132,7 +169,6 @@ export class ConfigPrompter {
      */
     constructor(private readonly systemLookup: SystemLookup, layer: FlexLayer, private readonly logger: ToolsLogger) {
         this.isCustomerBase = layer === FlexLayer.CUSTOMER_BASE;
-        this.ui5Info = UI5VersionInfo.getInstance(layer);
     }
 
     /**
@@ -360,7 +396,12 @@ export class ConfigPrompter {
 
         try {
             this.abapProvider = await getConfiguredProvider(options, this.logger);
-            await this.getSystemData();
+            const validationResult = await this.handleSystemDataValidation();
+
+            if (typeof validationResult === 'string') {
+                return validationResult;
+            }
+
             this.targetApps = await loadApps(this.abapProvider, this.isCustomerBase);
             this.isLoginSuccessful = true;
             return true;
@@ -414,7 +455,7 @@ export class ConfigPrompter {
     /**
      * Validates the selected application to ensure it is supported.
      *
-     * @param {Application} app - The application to validate.
+     * @param {SourceApplication} app - The application to validate.
      * @returns {Promise<boolean | string>} True if the application is valid, otherwise an error message.
      */
     private async validateAppData(app: SourceApplication): Promise<boolean | string> {
@@ -440,13 +481,12 @@ export class ConfigPrompter {
     /**
      * Evaluate if the application version supports certain features.
      *
-     * @param {Application} application - The application data.
+     * @param {SourceApplication} application - The application data.
      */
     private evaluateAppSupport(application: SourceApplication): void {
-        const systemVersion = this.ui5Info.systemVersion;
-        const isFullSupport = this.ui5Info.isVersionDetected && !isFeatureSupportedVersion('1.96.0', systemVersion);
+        const isFullSupport = !!this.systemVersion && !isFeatureSupportedVersion('1.96.0', this.systemVersion);
         const isPartialSupport =
-            this.ui5Info.isVersionDetected && isFullSupport && isFeatureSupportedVersion('1.90.0', systemVersion);
+            !!this.systemVersion && isFullSupport && isFeatureSupportedVersion('1.90.0', this.systemVersion);
 
         this.setSupportFlags(application, isFullSupport, isPartialSupport);
     }
@@ -456,7 +496,7 @@ export class ConfigPrompter {
      *
      * @returns A promise that resolves when system data is fetched.
      */
-    private async getSystemData(): Promise<void> {
+    private async loadSystemData(): Promise<void> {
         try {
             this.isCloudProject = await this.abapProvider.isAbapCloud();
             this.flexUISystem = await getFlexUISupportedSystem(this.abapProvider, this.isCustomerBase);
@@ -466,14 +506,26 @@ export class ConfigPrompter {
     }
 
     /**
+     * Fetches and processes SAPUI5 version data from the system and public sources.
+     *
+     * @returns {Promise<void>} A promise that resolves once all version data is loaded and assigned.
+     */
+    private async loadUI5Versions(): Promise<void> {
+        const version = await getSystemUI5Version(this.abapProvider);
+        this.systemVersion = checkSystemVersionPattern(version);
+        this.publicVersions = await fetchPublicVersions();
+        this.ui5Versions = await getRelevantVersions(this.systemVersion, this.isCustomerBase, this.publicVersions);
+    }
+
+    /**
      * Handles the fetching and validation of system data.
      *
      * @returns {Promise<boolean | string>} True if successful, or an error message if an error occurs.
      */
     private async handleSystemDataValidation(): Promise<boolean | string> {
         try {
-            await this.getSystemData();
-            await this.validateSystemVersion();
+            await this.loadSystemData();
+            await this.loadUI5Versions();
 
             if (!this.isCustomerBase && this.isCloudProject) {
                 return t('error.cloudSystemsForInternalUsers');
@@ -483,22 +535,6 @@ export class ConfigPrompter {
         } catch (e) {
             this.logger.debug(`Validating system failed. Reason: ${e.message}`);
             return e.message;
-        }
-    }
-
-    /**
-     * Validates the UI5 system version based on the provided value or fetches all relevant versions if no value is provided.
-     * Updates the internal state with the fetched versions and the detection status.
-     *
-     * @returns {Promise<void>} Resolves after checking system ui5 version.
-     */
-    private async validateSystemVersion(): Promise<void> {
-        try {
-            const version = await getSystemUI5Version(this.abapProvider);
-            await this.ui5Info.getRelevantVersions(version);
-        } catch (e) {
-            this.logger.debug(`Could not fetch system version: ${e.message}`);
-            await this.ui5Info.getRelevantVersions();
         }
     }